requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
graphene-django-cruddals-v1 @ git+ssh://git@github.com/juanjcardona13/graphene-django-cruddals-v1@main
decorator==5.1.1
Django==3.2.18
django-cors-headers==3.14.0
django-graphql-jwt==0.3.4
django-nested-admin==4.0.2
executing==1.2.0
graphene==3.2.2
graphene-django==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
promise==2.3
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.6.0
pypng==0.20220715.0
python-monkey-business==1.0.0
python-slugify==8.0.1
pytz==2023.3
qrcode==7.4.2
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
termcolor==2.3.0
text-unidecode==1.3
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
